fix find_element & controllato random_index

+++ src/pid_list/pid_list.c
@@ -33,13 +33,17 @@ void free_list(int* l) {
 
 int random_element(int* list, int size) {
   int* cpy = malloc(sizeof(int) * size);
-  int random_el = 0;
-  int found;
-  register int dim = size;
-  register int i = 0;
-  while (i < dim) {
-    cpy[i] = list[i];
-    i++;
+  if (cpy == NULL) {
+    return -1;
+  }
+  int random_el = 0, found = 0, dim = size;
+
+  {
+    register i = 0;
+    while (i < dim) {
+      cpy[i] = list[i];
+      i++;
+    }
   }
 
   /* ogni volta che il ciclo trova un processo terminato, lo separa dai processi attivi, spostando in fondo alla lista.
@@ -47,14 +51,16 @@ int random_element(int* list, int size) {
   do {
     int r_index = 0;
     srand(clock());
-    found = 1; /* imposto found a true */
     r_index = (rand() % dim);
-    random_el = cpy[r_index];
-    if (getpid() == random_el || (kill(random_el, 0) == -1 && errno == ESRCH)) {
-      found = 0;
-      cpy[r_index] = cpy[dim - 1];
-      cpy[dim - 1] = random_el;
-      dim--;
+    if (r_index >= 0 && r_index < dim) {
+      found = 1; /* imposto found a true */
+      random_el = cpy[r_index];
+      if (getpid() == random_el || (kill(random_el, 0) == -1 && errno == ESRCH)) {
+        found = 0;
+        dim--;
+        cpy[r_index] = cpy[dim];
+        cpy[dim] = random_el;
+      }
     }
   } while (dim > 0 && !found); /* evita di estrarre lo stesso processo in cui ci troviamo o di trovare un utente terminato*/
 
@@ -66,7 +72,7 @@ int find_element(int* l, int size, int pid) {
   int* ptr = l;
   int index = -1;
   int dim = size;
-  while (--dim > 0 && index != -1) {
+  while (--dim > 0 && index == -1) {
     if (ptr[dim] == pid) {
       index = dim;
     }
